requirements.txt
openai
fastapi[standard]
psycopg2-binary
pydantic
python-dotenv
requests
sqlalchemy
uvicorn[standard]
